requirements.txt
Django==3.1.4
djangorestframework==3.9.2
pytz==2018.9
sqlparse==0.3.0
django-filter
django-admin-relation-links
django-seed
django rest knox
